feat(cwl): Initialize TailLogGroup command with Wizard (#5722)
CWL is planning on supporting a LiveTail experience in AWSToolkit for
VSCode

Registers `aws.cwl.tailLogGroup` as a recognized command in AWS Toolkit.
In this PR, running this command will simply take the user through a
Wizard to collect the configuration for the tailing session, and logs
it.

In follow up PRs, I will take this Wizard input and use it to call CWL
APIs and output streamed LogEvents to a TextDocument.

This command will be able to be executed in the command pallet, or by
pressing a "play button" icon when viewing LogGroups in the CloudWatch
Logs explorer menu.