feat(cwl): Create LiveTailSession object and registry. Adds MaxLine configuration (#5738)
1. On the client side, VSCode needs to be aware of some additional
context of a running LiveTail session, other than just the response from
the StartLiveTail API call (Abort Controller, Max Lines config,
TextDocument URI, etc).

2. Additionally, we want to be able to have multiple LiveTail sessions
running, and have a way to organize all of them.

1. Create a LiveTailSession class in AWSToolkit that will persist all of
the metadata and context the LiveTail feature needs to operate.
2. Create a LiveTailSessionRegistry that maps a vscode URI to a
LiveTailSession. The URI for a LiveTailSession is composed of the API
request elements. This will allow us to tell when a user is making a
duplicate request. This URI will also be used as the URI of the
TextDocument to display the session results.

* Adds VSCode preference config for LiveTail max events. This will be
used to cap the number of live tail events can be in the TextDocument at
a given time. When the limit is reached, the oldest log events will be
dropped in order to fit new events streaming in.
* Adds a missing String for "Tail Log Group"