feat(cwl): initialize tailLogGroup command. starts stream and prints logEvents to textDocument (#5790)
TailLogGroup currently is just logging the Users WizardResponse. It does
not start a LiveTail stream or print results to the VSCode editor

* Creates a `LiveTailSession` with the input params from the
TailLogGroupWizard
* Registers the LiveTailSession in the `LiveTailSessionRegistry`
* Iterates the Response Stream and prints new LogEvents to a
TextDocument
* If there are no open tabs for a given session's TextDocument, the
stream will be closed by triggering its AbortController and disposing
its client.

We will use a new URI Scheme for LiveTail text documents so we can
identify a LiveTail session vs a Cloudwatch SearchLogGroup/ViewLogStream
document. In the future, we will register LiveTail specific CodeLenses
that depend upon this URI scheme. Defining a new URI scheme for a Text
Document requires a Document Provider. In this use case, it is empty
given that the response handling that writes to the document lives
within the TailLogGroup command itself.