feat(cwl): Support autoscrolling live tail session's visible editors (#5857)
## Problem
New log events in a LiveTail session are added to the end of a
TextDocument. Session updates can contain multiple LogEvents, leading to
the text document growing quickly past the user's view. requiring them
to constantly be scrolling the document themselves.

## Solution
To maintain parity with CWL console experience (and the classic `tail
-f` experience), we want to auto scroll the customer's LiveTail
session's visible editors to the bottom.

This will only happen IF the end of file is already in view. Meaning, if
a User scrolls UP (causing the end of file to not be in view), auto
scrolling will not enable. However, if they scroll back down to EOF
themselves, it will re-enable.

Simply, if the end of file is in view, when new logEvents are added to
the document, the editor will be autoscrolled back to the end of the
file.