fix(cwl): set LogEventFilter when constructing LiveTailSession #6008
## Problem
LogEventFilter is an optional configuration parameter when building a
LiveTailSession. Currently, the Constructor is not actually setting this
property so it is always undefined. This means that if a user sets a
Filter pattern, it will not be applied.

## Solution
Set the property in LiveTailSession Constructor.
Add missing unit tests to validate that LiveTailSession config values
are set properly. Testing via `buildStartLiveTailCommand` because this
is what actually gets sent to the CWL backend to start a
LiveTailSession.