fix(cwl): Dispose event listeners when command exits (#6095)
## Problem
TailLogGroup command is not being fully cleaned up properly. The event
listeners for:
* showing the StatusBar item
*  closing a session when exiting its editors  

are not being disposed of. As users stop sessions, these listeners
remain active whilst doing no meaningful work. This is effectively a
memory leak.

Additionally during testing, I noticed that we are adding a session to
the LiveTailRegistry before the session is actually started. This could
cause an issue where a session stream isn't successfully created, but
the session is still registered.

## Solution
* Create an array of disposables and dispose of them as we exit the
TailLogGroup command
* Only register a session after startLiveTail returns


Because now we dispose of the "tab close" listener when tailLogGroup
exits, changes are needed in the unit tests to keep the mock response
stream "open" as the test executes. This keeps tailLogGroup from
returning/disposing while we perform test assertions and close the
editors.