build(lint): disallow node process api, favor Toolkit ChildProcess api #6113
## Problem
contributors may use the low-level node `child_process` library, which
lacks the instrumentation and robustness of the toolkit `ChildProcess`
utility.
Example:
https://github.com/aws/aws-toolkit-vscode/pull/5925#discussion_r1841277911

## Solution
- add the node `child_process` library as a "banned import" with a
message pointing to our `ChildProcess` utility.
- Migrate simple cases to use our `ChildProcess`, mark more complex ones
as exceptions.

## Future Work
- To migrate the existing cases marked as exceptions, we will need to
add functionality to our `ChildProcess` api, such as enforcing a
`maxBufferSize` on the output.
- Additionally, some of the current uses of `child_process` are tied to
implementation details in the `child_process` library that make it hard
to change without a larger-scale refactor. These cases were marked as
exceptions.